refactor: extract shared modal focus-trap hook
ConfirmModal and HangmanGameOverModal carried identical 45-line
focus/Escape/Tab-cycling effects. Both now use useFocusTrap(dialogRef,
onEscape).
